feat: mvc구현, view, domain 분리
