+++ src/test/java/fifth_refactoring/CheckValidationTest.java
@@ -1,5 +1,6 @@
 package fifth_refactoring;
 
+import fifth_refactoring.view.CheckValidation;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.ValueSource;
@@ -7,14 +8,14 @@ import org.junit.jupiter.params.provider.ValueSource;
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
 
-public class ValidationCheckTest {
+public class CheckValidationTest {
 
     private CheckValidation checkValidation = new CheckValidation();
 
     @ParameterizedTest
     @ValueSource( ints = {-1 , 0})
     @DisplayName("1이하 벨리데이션체크")
-    void valdationTryNumberErrorTest(int testInput){
+    void validateTryNumberErrorTest(int testInput){
         assertThatThrownBy(()  -> {
             checkValidation.validateTryNumber(testInput);
         }).isInstanceOf(IllegalArgumentException.class).hasMessageContaining("1 이상의 숫자를 입력하세요");
@@ -23,14 +24,14 @@ public class ValidationCheckTest {
     @ParameterizedTest
     @ValueSource( ints = {1, 3})
     @DisplayName("정상값 벨리데이션체크")
-    void valdationTryNumberNormalTest(int testInput){
+    void validateTryNumberNormalTest(int testInput){
         assertThat(checkValidation.validateTryNumber(testInput)).isEqualTo(true);
     }
 
     @ParameterizedTest
     @ValueSource(strings = {" "})
     @DisplayName("빈값테스트")
-    void valdationCarNameTest(String input) {
+    void validateCarNameTest(String input) {
         assertThatThrownBy(()  -> {
             checkValidation.validateCarName(input);
         }).isInstanceOf(IllegalArgumentException.class).hasMessageContaining("공백금지");
