+++ src/test/java/fifth_refactoring/InputViewTest.java
@@ -1,5 +1,6 @@
 package fifth_refactoring;
 
+import fifth_refactoring.view.InputView;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.ValueSource;
