+++ src/test/java/fifth_refactoring/ResultViewTest.java
@@ -1,9 +1,8 @@
 package fifth_refactoring;
 
+import fifth_refactoring.domain.Car;
 import org.junit.jupiter.api.Test;
-import fifth_refactoring.Car;
-import fifth_refactoring.RacingGame;
-import fifth_refactoring.ResultView;
+import fifth_refactoring.view.ResultView;
 
 import java.util.ArrayList;
 import java.util.List;
@@ -22,5 +21,21 @@ public class ResultViewTest {
         car.move();
         car.move();
         assertThat(resultview.makeSign(car.getCarPosition())).isEqualTo("---");
+    }
+
+    @Test
+    void setRankTest() {
+        List<Car> carList = new ArrayList<Car>();
+
+        Car car1 = new Car();
+        car1.setCarName("test1");
+        carList.add(car1);
+
+        Car car2 = new Car();
+        car2.setCarName("test2");
+        car2.move();
+        carList.add(car2);
+
+        assertThat(ResultView.setRank(carList)).isEqualTo("test2");
     }
 }
